#352 New partial for attachmentshow
