+++ resources/js/pages/Attachments/Show.vue
@@ -1,11 +1,11 @@
 <script setup lang="ts">
 import { Head, Link, router } from '@inertiajs/vue3'
 import AppLayout from '@/layouts/app/AppSidebarLayout.vue'
-import { ref, computed, onMounted } from 'vue'
+import { ref, onMounted } from 'vue'
 import { type BreadcrumbItem } from '@/types'
 import { route } from 'ziggy-js'
-import { deleteAttachments } from '@/services/attachmentService'
-import { fetchAttachment } from '@/services/attachmentService'
+import { deleteAttachments, fetchAttachment } from '@/services/attachmentService'
+import AttachmentDetailSection from './components/AttachmentDetailSection.vue'
 
 interface UserPermissions {
     view: boolean
@@ -50,7 +50,7 @@ const attachment = ref<Attachment>({
     attachable: props.attachment.attachable,
     updated_at: props.attachment.updated_at,
     is_test: props.attachment.is_test ?? false,
-    permissions: { view: false, update: false, delete: false },
+    permissions: props.attachment.permissions ?? { view: false, update: false, delete: false },
 })
 
 const breadcrumbItems: BreadcrumbItem[] = [
@@ -58,24 +58,13 @@ const breadcrumbItems: BreadcrumbItem[] = [
     { title: props.attachment.filename, href: route('attachments.show', { attachment: attachment.value.id }) },
 ]
 
-function capitalize(str: string | null | undefined) {
-    if (!str) return '-'
-    return str
-        .split(' ')
-        .map(s => s.charAt(0).toUpperCase() + s.slice(1))
-        .join(' ')
-}
-
 async function loadAttachment() {
     const data = await fetchAttachment(attachment.value.id)
-
     Object.assign(attachment.value, data)
 }
+
 const showDeleteModal = ref(false)
 const deleting = ref(false)
-
-const isImage = computed(() => props.attachment.mime?.startsWith('image/'))
-const isPdf = computed(() => props.attachment.mime === 'application/pdf')
 
 async function handleDelete() {
     deleting.value = true
@@ -90,27 +79,15 @@ async function handleDelete() {
     }
 }
 
-function formatSize(bytes: number | null): string {
-    if (!bytes) return '—'
-    if (bytes < 1024) return bytes + ' B'
-    if (bytes < 1024 * 1024) return (bytes / 1024).toFixed(1) + ' KB'
-    return (bytes / (1024 * 1024)).toFixed(1) + ' MB'
+function capitalize(str: string | null | undefined) {
+    if (!str) return '—'
+    return str
+        .split(' ')
+        .map(s => s.charAt(0).toUpperCase() + s.slice(1))
+        .join(' ')
 }
 
-function formatDate(dateString: string | null): string {
-    if (!dateString) return '—'
-    return new Date(dateString).toLocaleDateString('en-GB', {
-        day: '2-digit',
-        month: 'short',
-        year: 'numeric',
-        hour: '2-digit',
-        minute: '2-digit',
-    })
-}
-
-onMounted(() => {
-    loadAttachment()
-})
+onMounted(() => loadAttachment())
 </script>
 
 <template>
@@ -118,7 +95,6 @@ onMounted(() => {
         <Head :title="capitalize(attachment.filename) || 'Attachment'" />
 
         <div class="p-6 max-w-4xl space-y-6">
-
             <!-- Header actions -->
             <div class="flex items-center justify-between">
                 <h1 class="text-2xl font-bold">
@@ -150,78 +126,7 @@ onMounted(() => {
                 </div>
             </div>
 
-            <!-- File Preview -->
-            <div class="border rounded p-6">
-                <h2 class="text-sm font-medium text-gray-500 mb-4">Preview</h2>
-
-                <img
-                    v-if="isImage"
-                    :src="attachment.preview_url ?? attachment.download_url"
-                    :alt="attachment.filename"
-                    class="max-w-full max-h-96 rounded border object-contain"
-                />
-
-                <iframe
-                    v-else-if="isPdf"
-                    :src="attachment.preview_url ?? attachment.download_url"
-                    class="w-full h-96 border rounded"
-                ></iframe>
-
-                <div
-                    v-else
-                    class="flex flex-col items-center justify-center py-12 rounded border"
-                >
-                    <svg class="w-16 h-16 text-gray-300" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                        <path stroke-linecap="round" stroke-linejoin="round" stroke-width="1.5" d="M9 12h6m-6 4h6m2 5H7a2 2 0 01-2-2V5a2 2 0 012-2h5.586a1 1 0 01.707.293l5.414 5.414a1 1 0 01.293.707V19a2 2 0 01-2 2z" />
-                    </svg>
-                    <p class="mt-2 text-sm text-gray-500">No preview available</p>
-                    <a
-                        :href="attachment.download_url"
-                        class="px-4 py-2 text-sm border border-gray-300 rounded"
-                        :class="attachment.is_test ? 'opacity-50 pointer-events-none cursor-not-allowed' : 'hover:bg-gray-50'"
-                        target="_blank"
-                        :aria-disabled="attachment.is_test"
-                    >
-                        Download to view
-                    </a>
-                </div>
-            </div>
-
-            <!-- Details -->
-            <div class="border rounded p-6">
-                <h2 class="text-sm font-medium text-gray-500 mb-4">Details</h2>
-
-                <dl class="grid grid-cols-1 gap-4 sm:grid-cols-2">
-                    <div>
-                        <dt class="text-xs font-medium text-gray-500 uppercase tracking-wide">Filename</dt>
-                        <dd class="mt-1 text-sm">{{ attachment.filename }}</dd>
-                    </div>
-                    <div>
-                        <dt class="text-xs font-medium text-gray-500 uppercase tracking-wide">File Size</dt>
-                        <dd class="mt-1 text-sm">{{ formatSize(attachment.size) }}</dd>
-                    </div>
-                    <div>
-                        <dt class="text-xs font-medium text-gray-500 uppercase tracking-wide">File Type</dt>
-                        <dd class="mt-1 text-sm">{{ attachment.mime || '—' }}</dd>
-                    </div>
-                    <div>
-                        <dt class="text-xs font-medium text-gray-500 uppercase tracking-wide">Attached To</dt>
-                        <dd class="mt-1 text-sm">{{ capitalize(attachment.attachable_type) }}</dd>
-                    </div>
-                    <div>
-                        <dt class="text-xs font-medium text-gray-500 uppercase tracking-wide">Uploaded By</dt>
-                        <dd class="mt-1 text-sm">{{ attachment.uploader?.name || '-' }}</dd>
-                    </div>
-                    <div>
-                        <dt class="text-xs font-medium text-gray-500 uppercase tracking-wide">Uploaded At</dt>
-                        <dd class="mt-1 text-sm">{{ formatDate(attachment.created_at) }}</dd>
-                    </div>
-                    <div v-if="attachment.updated_at !== attachment.created_at">
-                        <dt class="text-xs font-medium text-gray-500 uppercase tracking-wide">Last Updated</dt>
-                        <dd class="mt-1 text-sm">{{ formatDate(attachment.updated_at) }}</dd>
-                    </div>
-                </dl>
-            </div>
+            <AttachmentDetailSection :attachment="attachment" />
 
             <!-- Danger Zone -->
             <div v-if="attachment.permissions.delete" class="border border-red-200 rounded p-6">
@@ -242,7 +147,7 @@ onMounted(() => {
             class="fixed inset-0 bg-black/50 flex items-center justify-center z-50"
             @click.self="showDeleteModal = false"
         >
-            <div class="rounded-lg shadow-xl p-6 max-w-md w-full mx-4">
+            <div class="bg-white rounded-lg shadow-xl p-6 max-w-md w-full mx-4">
                 <h2 class="text-lg font-medium mb-2">Delete attachment?</h2>
                 <p class="text-sm text-gray-600 mb-6">
                     This will permanently remove
